ci: fix code issues of golangci-lint on Windows
